website_builder_1/requirements.txt
wget
opencv-python-headless
dlib
numpy
pillow
beautifulsoup4
requests
python-dotenv 